simplified file structure
